+++ NEWS.md
@@ -3,17 +3,17 @@
 This is a **major refactoring/rewrite** of checkpoint, aimed at solving many long-standing issues.
 
 - Switch to [pkgdepends](https://github.com/r-lib/pkgdepends) for the backend, replacing custom-written code that calls `install.packages` and the like. This brings the following benefits:
-  - Support for parallel installs.
   - Caching of downloaded packages.
   - Allow installing packages which are in use, without having to unload them first.
   - Comprehensive reporting of all aspects of the install process: dependency resolution, creating an install plan, downloading packages, and actual installation.
   - Reliable detection of installation outcomes (no more having to screen-scrape the R terminal!)
 - New functions `create_checkpoint` and `use_checkpoint`, reflecting the two main objectives of the package. The `checkpoint` function calls out to these, based on whether the checkpoint directory exists.
   - A side-effect of this is that calling `checkpoint` in your home directory should result in checkpointing very many packages only once.
-- New functions `uncheckpoint_session` (which is the reverse of `use_checkpoint`), `delete_checkpoint` and `delete_all_checkpoints` to manage checkpointing. They replace `checkpointRemove`, `checkpointArchives` and `unCheckpoint`.
+- New functions `uncheckpoint` (which is the reverse of `use_checkpoint`), `delete_checkpoint` and `delete_all_checkpoints` to manage checkpointing. They replace `checkpointRemove`, `checkpointArchives` and `unCheckpoint`.
 - Function `getValidSnapshots` renamed to `list_mran_snapshots` to clarify that it lists the snapshots on the MRAN server. Similarly, `setSnapshot` renamed to `use_mran_snapshot` for consistency with other function names.
 - Function `scanForPackages` renamed to `scan_project_files` to match grammatical pattern of other function names; now automatically includes rmarkdown (not knitr) in the list of dependencies if Rmarkdown-based files are found.
 - Experimental support for a `checkpoint.yml` manifest file, to specify packages to include or exclude.
+   - Allows including packages from Bioconductor or GitHub in the checkpoint; it's recommended to specify a version number, tag or commit hash to ensure you always get the same version of a package.
 - Consistent use of snake_case for all object names and function arguments.
 - Remove obsolete code to handle lack of HTTPS support in ancient versions of R. (Note that the MRAN site now requires HTTPS.)
 - Other bug fixes and general tidying.